requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow==2.15.0
numpy
Pillow
opencv-python-headless